src/main/java/pe/edu/unfv/emplos/models/Cuenta.java
package pe.edu.unfv.emplos.models;

import pe.edu.unfv.emplos.exceptions.DineroInsuficienteException;

import java.math.BigDecimal;
import java.util.Objects;

public class Cuenta {

    private String persona;
    private BigDecimal saldo;

    public Cuenta(String persona, BigDecimal saldo) {
        this.persona = persona;
        this.saldo = saldo;
    }

    public String getPersona() {
        return persona;
    }

    public void setPersona(String persona) {
        this.persona = persona;
    }

    public BigDecimal getSaldo() {
        return saldo;
    }

    public void setSaldo(BigDecimal saldo) {
        this.saldo = saldo;
    }

    void debito(BigDecimal valor) {
        BigDecimal nuevoSaldo = this.saldo.subtract(valor);
        if (nuevoSaldo.compareTo(BigDecimal.ZERO) < 0) throw new DineroInsuficienteException("Dinero insuficiente");

        this.saldo = nuevoSaldo;
    }
    void credito(BigDecimal valor) {
        this.saldo = this.saldo.add(valor);
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Cuenta cuenta = (Cuenta) o;
        return Objects.equals(persona, cuenta.persona) && Objects.equals(saldo, cuenta.saldo);
    }

    @Override
    public int hashCode() {
        return Objects.hash(persona, saldo);
    }
}
